feat: criado o componente rowsTable
