+++ frontend/src/pages/users/index.tsx
@@ -1,9 +1,11 @@
 import Header from "../../components/header";
+import TableInfo from "../../components/table";
 
 const Users = () => {
   return (
     <>
       <Header />
+      <TableInfo/>
     </>
   );
 };
